Rename analysis engine -> language server
