+++ src/client/activation/languageServerHashes.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+// This file will be replaced by a generated one during the release build
+// with actual hashes of the uploaded packages.
+// Values are for test purposes only
+export const language_server_win_x86_sha512 = 'win-x86';
+export const language_server_win_x64_sha512 = 'win-x64';
+export const language_server_osx_x64_sha512 = 'osx-x64';
+export const language_server_linux_x64_sha512 = 'linux-x64';
