+++ src/client/telemetry/constants.ts
@@ -32,8 +32,8 @@ export const UNITTEST_STOP = 'UNITTEST.STOP';
 export const UNITTEST_RUN = 'UNITTEST.RUN';
 export const UNITTEST_DISCOVER = 'UNITTEST.DISCOVER';
 export const UNITTEST_VIEW_OUTPUT = 'UNITTEST.VIEW_OUTPUT';
-export const PYTHON_ANALYSIS_ENGINE_ENABLED = 'PYTHON_ANALYSIS_ENGINE.ENABLED';
-export const PYTHON_ANALYSIS_ENGINE_DOWNLOADED = 'PYTHON_ANALYSIS_ENGINE.DOWNLOADED';
-export const PYTHON_ANALYSIS_ENGINE_ERROR = 'PYTHON_ANALYSIS_ENGINE.ERROR';
-export const PYTHON_ANALYSIS_ENGINE_STARTUP = 'PYTHON_ANALYSIS_ENGINE.STARTUP';
+export const PYTHON_LANGUAGE_SERVER_ENABLED = 'PYTHON_LANGUAGE_SERVER.ENABLED';
+export const PYTHON_LANGUAGE_SERVER_DOWNLOADED = 'PYTHON_LANGUAGE_SERVER.DOWNLOADED';
+export const PYTHON_LANGUAGE_SERVER_ERROR = 'PYTHON_LANGUAGE_SERVER.ERROR';
+export const PYTHON_LANGUAGE_SERVER_STARTUP = 'PYTHON_LANGUAGE_SERVER.STARTUP';
 export const TERMINAL_CREATE = 'TERMINAL.CREATE';
