+++ src/test/definitions/hover.ptvs.test.ts
@@ -5,7 +5,7 @@ import * as assert from 'assert';
 import * as path from 'path';
 import * as vscode from 'vscode';
 import '../../client/common/extensions';
-import { IsAnalysisEngineTest } from '../constants';
+import { IsLanguageServerTest } from '../constants';
 import { closeActiveWindows, initialize, initializeTest } from '../initialize';
 import { normalizeMarkedString } from '../textUtils';
 
@@ -21,9 +21,9 @@ const fileStringFormat = path.join(hoverPath, 'stringFormat.py');
 let textDocument: vscode.TextDocument;
 
 // tslint:disable-next-line:max-func-body-length
-suite('Hover Definition (Analysis Engine)', () => {
+suite('Hover Definition (Language Server)', () => {
     suiteSetup(async function () {
-        if (!IsAnalysisEngineTest()) {
+        if (!IsLanguageServerTest()) {
             // tslint:disable-next-line:no-invalid-this
             this.skip();
         }
