+++ src/test/definitions/parallel.ptvs.test.ts
@@ -6,20 +6,20 @@ import { EOL } from 'os';
 import * as path from 'path';
 import * as vscode from 'vscode';
 import { IS_WINDOWS } from '../../client/common/platform/constants';
-import { IsAnalysisEngineTest } from '../constants';
+import { IsLanguageServerTest } from '../constants';
 import { closeActiveWindows, initialize } from '../initialize';
 import { normalizeMarkedString } from '../textUtils';
 
 const autoCompPath = path.join(__dirname, '..', '..', '..', 'src', 'test', 'pythonFiles', 'autocomp');
 const fileOne = path.join(autoCompPath, 'one.py');
 
-suite('Code, Hover Definition and Intellisense (MS Python Code Analysis)', () => {
+suite('Code, Hover Definition and Intellisense (Language Server)', () => {
     suiteSetup(async function () {
         // https://github.com/Microsoft/vscode-python/issues/1061
         // tslint:disable-next-line:no-invalid-this
         this.skip();
 
-        if (!IsAnalysisEngineTest()) {
+        if (!IsLanguageServerTest()) {
             // tslint:disable-next-line:no-invalid-this
             this.skip();
         }
