+++ src/test/performanceTest.ts
@@ -41,10 +41,10 @@ class TestRunner {
         const timesToLoadEachVersion = 2;
         const devLogFiles: string[] = [];
         const releaseLogFiles: string[] = [];
-        const newAnalysisEngineLogFiles: string[] = [];
+        const languageServerLogFiles: string[] = [];
 
         for (let i = 0; i < timesToLoadEachVersion; i += 1) {
-            await this.enableNewAnalysisEngine(false);
+            await this.enableLanguageServer(false);
 
             const devLogFile = path.join(logFilesPath, `dev_loadtimes${i}.txt`);
             console.log(`Start Performance Tests: Counter ${i}, for Dev version with Jedi`);
@@ -56,18 +56,18 @@ class TestRunner {
             await this.capturePerfTimes(Version.Release, releaseLogFile);
             releaseLogFiles.push(releaseLogFile);
 
-            // New Analysis engine.
-            await this.enableNewAnalysisEngine(true);
-            const newAnalysisEngineLogFile = path.join(logFilesPath, `newAnalysisEngine_loadtimes${i}.txt`);
-            console.log(`Start Performance Tests: Counter ${i}, for Release version with Analysis Engine`);
-            await this.capturePerfTimes(Version.Release, newAnalysisEngineLogFile);
-            newAnalysisEngineLogFiles.push(newAnalysisEngineLogFile);
+            // Language server.
+            await this.enableLanguageServer(true);
+            const languageServerLogFile = path.join(logFilesPath, `languageServer_loadtimes${i}.txt`);
+            console.log(`Start Performance Tests: Counter ${i}, for Release version with Language Server`);
+            await this.capturePerfTimes(Version.Release, languageServerLogFile);
+            languageServerLogFiles.push(languageServerLogFile);
         }
 
         console.log('Compare Performance Results');
-        await this.runPerfTest(devLogFiles, releaseLogFiles, newAnalysisEngineLogFiles);
+        await this.runPerfTest(devLogFiles, releaseLogFiles, languageServerLogFiles);
     }
-    private async enableNewAnalysisEngine(enable: boolean) {
+    private async enableLanguageServer(enable: boolean) {
         const settings = `{ "python.jediEnabled": ${!enable} }`;
         await fs.writeFile(path.join(EXTENSION_ROOT_DIR, 'src', 'test', 'performance', 'settings.json'), settings);
     }
